fix: address PR review, clippy, audit, and formatting

+++ crates/engine-processing/src/io/source/db_reader.rs
@@ -323,7 +323,7 @@ impl DbSourceReader {
             .columns(columns)
             .joins(extra_joins)
             .filter(filter_clause)
-            .limit(i64::MAX as usize)
+            .limit(usize::MAX)
             .cursor(Cursor::None)
             .strategy(OffsetStrategyFactory::default_strategy());
